LWSHADOOP-749: Explicitly declare jackson dependency
Certain classes in `solr-hadoop-common` (specifically, LWSolrDocument)
make use Jackson APIs, but nothing in the project's gradle file declares
its reliance on this dependency.  This means that we're implicitly
relying on one of our other dependencies (hadoop) to pull it in transitively.

This is misleading and trappy.  It also causes opens the door to build
failures, as we allow our consumers to pick their own hadoop version.

This commit declares our dependence on Jackson in our build.gradle.