LWSHADOOP-565: Make Hadoop config profiles optional
Prior to this commit, solr-hadoop-common consumers were required to
setup both a hadoop2Version and a hadoop3Version dependency configuration.

This made sense in the context of the hadoop-solr job-jar, which drove
these changes, but doesn't make sense for all consumers.  Some parent
modules will need separate branches to support Hadoop 2/3.  Others may
just not be interested in building/testing both simultaneously.

This commit alters the dependency declarations in solr-hadoop-common
to only be invoked if the 'hadoop2Version' or 'hadoop3Version' is
defined, respectively.  This gives solr-hadoop-common the flexibility
to be consumed by projects that just care about just one of the two, or
about both.