LWSHADOOP-703: Shade solr-test-framework
Some solr-hadoop-common consumers run tests that utilize Hadoop
in-memory clusters (for example, hadoop-minicluster).  As our packaging
is currently setup, this is potentially troublesome: particularly if
these other test servers rely on Jetty versions differing from that
pulled in by Solr.

To make this easier on consumers, this commit adds a new gradle
submodule: 'solr-hadoop-shaded-test-framework', whose entire job is to
produce a shaded solr-test-framework uber jar.  This uber jar shades its
Jetty dependency, removing it as a potential cause of
classpath-collision.