LWSHADOOP-775: Ensure input-spliting handles aliases
solr-hadoop-common creates an input split for each shard in the Solr
input.  This usually works well, but when the "collection" provided is
an alias, we fail to extract its shards and end up creating a single
input split that does a distrib=false query on the alias itself.  When
aliases represent more than one collection, this results in only some of
the documents being returned.

This commit redoes our input-split logic to handle aliases differently.
When an alias is encountered, we'll now look up the collections that
belong to it and collect all the shards for all of those collections
together as splits.